refactor(core): home stream window tests
Move SliceWindows/WindowSpec tests and FileWindows geometry tests
out of stream.rs into stream/{tests,file_windows_tests}.rs.
